code/requirements.txt
fastapi
uvicorn
requests
python-multipart 
psycopg2-binary
numpy
pydantic
PyMuPDF
pymongo
dnspython
python-dotenv
